+++ libs/croissant-gl/src/public/object.ts
@@ -1,6 +1,6 @@
 import {DrawableType} from "../lib/graphics/drawable-type";
-import {objectBroker} from "../lib/object-broker";
-import {objectPropertiesBroker} from "../lib/object-properties-broker";
+import {objectBroker} from "../lib/brokers/object-broker";
+import {objectPropertiesBroker} from "../lib/brokers/object-properties-broker";
 import {renderer} from "../lib/renderer";
 import {ObjectInfo} from "../lib/types/object";
 import {vec3} from "gl-matrix";
